+++ source/extensions/isaacsim.examples.interactive/isaacsim/examples/interactive/robo_party/robo_party_extension.py
@@ -21,8 +21,8 @@ import os
 
 import omni.ext
 import omni.ui as ui
+from isaacsim.examples.base.base_sample_extension_experimental import BaseSampleUITemplate
 from isaacsim.examples.browser import get_instance as get_browser_instance
-from isaacsim.examples.interactive.base_sample import BaseSampleUITemplate
 from isaacsim.examples.interactive.robo_party import RoboParty
 from isaacsim.gui.components.ui_utils import btn_builder
 
@@ -63,7 +63,6 @@ class RoboPartyExtension(omni.ext.IExt):
 
         get_browser_instance().register_example(
             name=self.example_name,
-            execute_entrypoint=ui_handle.build_window,
             ui_hook=ui_handle.build_ui,
             category=self.category,
         )